Move `IntegrationTypeSerializer` to CustomFields namespace
